% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_index.R
\name{calc_d_index}
\alias{calc_d_index}
\title{Calculate the refined index of agreement}
\usage{
calc_d_index(predicted, observed)
}
\arguments{
\item{predicted}{numeric vector containing predicted values}

\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}
}
\value{
numeric outcome of the refined index of agreement
}
\description{
This function calculates the refined index of agreement \insertCite{willmott_refined_2012}{weatherQC}, 
which is an evaluation metric especially suited for precipitation occurrence 
in aerid regions \insertCite{aguilera_estimating_2020}{weatherQC}.
}
\details{
The refined index of agreement can range from -1 to 1, where a value close to one indicates good
model performance to predict precipitation occurrence. It is said to approach
values of 1 slowly, so it is also suited to discriminated well performing
models.
}
\examples{
calc_d_index(predicted = c(rep(0,5), rep(1, 5)), 
observed = c(0,0, 0, 1, 1, 1, 1, 0, 0, 0))
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
