% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hanssen_kuipers.R
\name{calc_hanssen_kuipers}
\alias{calc_hanssen_kuipers}
\title{Calculate Hansse-Kuipers discriminant}
\usage{
calc_hanssen_kuipers(observed, predicted)
}
\arguments{
\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}

\item{predicted}{numeric vector containing predicted values}
}
\value{
numeric outcome of the Hansse - Kuipers discriminant
}
\description{
This function calculates Hanssen - Kuipers discriminant (HK) 
\insertCite{hanssen_relationship_1965}{weatherQC}, which is an 
evaluation metric especially suited for precipitation occurrence.
}
\details{
The HK can range from -1 to 1, where a value close to one indicates good
model performance to predict precipitation occurrence.
}
\examples{
calc_hanssen_kuipers(predicted = c(rep(0,5), rep(1, 5)), 
observed = c(0,0, 0, 1, 1, 1, 1, 0, 0, 0))
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
