% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hit_score.R
\name{calc_hit_score}
\alias{calc_hit_score}
\title{Calculate hit score}
\usage{
calc_hit_score(observed, predicted)
}
\arguments{
\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}

\item{predicted}{numeric vector containing predicted values}
}
\value{
numeric outcome of the hit score
}
\description{
This function calculates the hit score, which is an evaluation metric
especially suited for precipitation.
}
\details{
The hitscore is the mean of correctly identified precipitation events and 
correctly identified dry-events \insertCite{simolo_improving_2010,oriani_missing_2020}{weatherQC}.
The function does not evaluate the imputed precipitation amount but only the 
precipitation occurrence.
}
\examples{
calc_hit_score(predicted = c(rep(0,5), rep(1, 5)), 
observed = c(0,0, 0, 1, 1, 1, 1, 0, 0, 0))
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
