% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_agreement.R
\name{calc_index_agreement}
\alias{calc_index_agreement}
\title{Calculate index of agreement}
\usage{
calc_index_agreement(x, y)
}
\arguments{
\item{x}{= target observation}

\item{y}{= neighbor observation}
}
\value{
index of agreement
}
\description{
Function to calculate index of agreement
}
\details{
The index of agreement expresses the ratio of mean square error and the potential error.
It expresses both the covariation and the absolute difference of target and auxiliary
weather station. The index of agreement can range from 0 to 1, the higher the value
the better the agreement of the compared objects. The index of agreement is 
sensitive to extreme values.
}
\examples{
calc_index_agreement(x = 1:10, y = 2:11)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
