% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MAE.R
\name{calc_MAE}
\alias{calc_MAE}
\title{Calculate mean absolute error}
\usage{
calc_MAE(predicted, observed, na.rm = FALSE)
}
\arguments{
\item{predicted}{numeric vector containing predicted values}

\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}

\item{Boolean}{parameter indicating whether NA values should be removed before the analysis}
}
\value{
numeric value of mean absolute error
}
\description{
This function calculates the mean absolute error for predicted and 
observed values.
}
\details{
The lower the MAE the better, it ranges from 0 to positive infinity. It is
interpreted to be the mean distance of prediction to observation, with 
equal weight for each residual. It is said to perform better if the error
is uniformly distributed.
}
\examples{
calc_MAE(predicted = 1:10, observed = 2:11)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
