% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MCC.R
\name{calc_MCC}
\alias{calc_MCC}
\title{Calculate Matthews Correlation Coefficient}
\usage{
calc_MCC(observed, predicted)
}
\arguments{
\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}

\item{predicted}{numeric vector containing predicted values}
}
\value{
numeric outcome of the MCC
}
\description{
This function calculates Matthews correlation coefficient (MCC), which is an evaluation metric
especially suited for unbalanced classification problems (like precipitation
occurrence)
}
\details{
The MCC can range from -1 to 1, where a value close to one indicates good
performance. It handles imbalanced groups better than other popular
classification problem metrics like F1 or accuracy score.
}
\examples{
calc_MCC(predicted = c(rep(0,5), rep(1, 5)), 
observed = c(0,0, 0, 1, 1, 1, 1, 0, 0, 0))
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
