% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_NSE.R
\name{calc_NSE}
\alias{calc_NSE}
\title{Calculate mean Nash - Sutcliffe Modelleing Efficiency Coefficient (NSE)}
\usage{
calc_NSE(predicted, observed)
}
\arguments{
\item{predicted}{numeric vector containing predicted values}

\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}

\item{Boolean}{parameter indicating whether NA values should be removed before the analysis}
}
\value{
numeric value of the NSE
}
\description{
This function calculates the Nash - Sutcliffe Modelling Efficiency Coefficient
(NSE) for predicted and observed values.
}
\details{
The higher the NSE, the better the model performance- THE NSE can range from
negative infinity to +1. It relates the magnitude of imputed data
variance compared to the variance in observed data 
\insertCite{nash_river_1970}{weatherQC}.
}
\examples{
calc_NSE(predicted = 1:10, observed = 2:11)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
