% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_S_index.R
\name{calc_S_index}
\alias{calc_S_index}
\title{Calculate S-index}
\usage{
calc_S_index(predicted, observed)
}
\arguments{
\item{predicted}{numeric vector containing predicted values}

\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}
}
\value{
numeric outcome of the S-index
}
\description{
This function calculates the S-index for predicted and observed data.
}
\details{
The S-index can range from 0 to 1, with higher values indicating better
model performance. It is said to express the agreement percentage 
of predicted and observed values. It can detect proportional difference
in observed and predicted means and is sensitive to extreme values.
}
\examples{
calc_S_index(predicted = 1:10, observed = 2:11)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
