% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_skill_score.R
\name{calc_skill_score}
\alias{calc_skill_score}
\title{Calculate skill score}
\usage{
calc_skill_score(predicted, observed)
}
\arguments{
\item{predicted}{numeric vector containing predicted values}

\item{observed}{numeric vector containing observed values, should be of same
length as \code{predicted}}
}
\value{
numeric outcome of the skill score
}
\description{
This function calculates the skill score for predicted and observed data.
}
\details{
The skill score can range from negative infinity to 1, with higher values 
indicating better model performance. A value lower than 0 indicates that on average the
difference of observation to its mean is smaller than the difference of predicted
value to observed value. Or to put it simply, always using the observed mean
instead of using the model results to less error. A value above 0 indicates
a benefit in using the model.
}
\examples{
calc_skill_score(predicted = 1:10, observed = 2:11)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
