% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clim_percentiles_prec.R
\name{get_clim_percentiles_prec}
\alias{get_clim_percentiles_prec}
\title{Calculate climatic percentile of precipitation}
\usage{
get_clim_percentiles_prec(
  weather,
  doy,
  probs = c(0.3, 0.5, 0.7, 0.9),
  min_non_zero_days = 20
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{doy}{Day of the year for which the empricial cumulative density function of precipitatio should be calculated}

\item{probs}{numeric vector, indicating the percent percentiles that sould be calculated}

\item{min_non_zero_days}{minimum amount of non-zero precipitation observation need 
to be present for ecdf calculation, otherwise NA returned}
}
\value{
vector of same length as probs, containing the precipitation values for 
the supplied percentiles
}
\description{
Function calculates percentile of non-zero precipitation for a certain day of the year.
}
\details{
This is a helper function which is mainly used for the spatial corrobation test
of precipitation. The function includes precipitation data of a 29 day window centered at
the day of interest throughout all observed years. Missing observations and 
zer-precipitation observations are discarded. If there are fewer observations than 
\code{min_non_zero_days}, then NA is returned instead of the empirical cumulative
density function.
}
\examples{
get_clim_percentiles_prec(weather = weather, doy = 1)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
