% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_y.R
\name{get_closest_y}
\alias{get_closest_y}
\title{Retrieves closest value}
\usage{
get_closest_y(x, y, i)
}
\arguments{
\item{x}{numerical value}

\item{y}{vector, for which the closest value compared to x is searched}
}
\value{
value in \code{y} which is the closest to \code{x}
}
\description{
Function finds closest value of y for x
}
\details{
This is a helper function needed for the spatial consistency test. It simply 
returns the closest value in y for the value of x. If there are two or more
values in y which are the closest to x, the first one is returned. In case x 
or y only contains NA, the function returns NA instead.
}
\examples{
get_closest_y(x = 1, y = c(0,2,3))
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
