% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicated_values.R
\name{get_duplicated_values}
\alias{get_duplicated_values}
\title{Identify instances of duplicated weather records}
\usage{
get_duplicated_values(
  weather,
  variable,
  precip_min_nonzero = 3,
  same_temp_threshold = 10
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{column name in \code{weather} for which the test is performed. Should
be either Tmin or Tmax. data.frames in \code{aux_list} need to have the same
name}

\item{precip_min_nonzero}{minimum amount of non-zero precipitation events for
a period to be considered in the test, otherwise the period is skipped}

\item{same_temp_threshold}{threshold for a month to contain same Tmin as Tmax
observation to be flagged.}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
The function performs several tests to identify duplicated months / years
within the weather data.
}
\details{
The function tests there are instances of duplicated years (Precipitation),
duplicated months within the same year (Temperature and precipitation),
duplicated months among the same month of different years (temperature and precipitation)
and cases of Tmin = Tmax. In case of detected duplicates, all observation
of the tested variable belonging to the duplicated period are flagged. In case of
precipitation a period needs to contain at least three non-zero precipitation
events in order to be included in the duplication tests. Otherwise these periods
are skipped.

This function is part of the weather quality control scheme after \insertCite{durre_comprehensive_2010;textual}{weatherQC}. For more details 
please refer to section 3 "Basic integrity tests".
}
\examples{
get_duplicated_values(weather = target_weather, variable = "Tmin")
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
