% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_each_day_precipitation_percentile.R
\name{get_each_day_precipitation_percentile}
\alias{get_each_day_precipitation_percentile}
\title{calculate percentiles for each day of the year}
\usage{
get_each_day_precipitation_percentile(
  weather,
  probs = c(0.3, 0.5, 0.7, 0.9),
  min_non_zero_days = 20
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{probs}{numeric vector, indicating the percent percentiles that sould be 
calculated}

\item{min_non_zero_days}{minimum amount of non-zero precipitation observation need 
to be present for ecdf calculation, otherwise NA returned}
}
\value{
data.frame with columns c("doy", probs) and 366 rows, one for each
day of the year (leap years are included). In the columns of probs the
respective percentiles are included. If the percentile calculation failed
because of the lack of non-zero data, then the row will contain NA instead
}
\description{
This function calculates climatological precipitation percentiles for each
day of the year.
}
\details{
This is a wrapper function of \code{\link{get_clim_percentiles_prec}}.
The function includes precipitation data of a 29 day window centered at
the day of interest throughout all observed years. Missing observations and 
zer-precipitation observations are discarded. If there are fewer observations than 
\code{min_non_zero_days}, then NA is returned instead of the empirical cumulative
density function.
}
\examples{
get_each_day_precipitation_percentile(weather = weather)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
