% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecdf.R
\name{get_ecdf}
\alias{get_ecdf}
\title{Helper function for spatial corroboration test}
\usage{
get_ecdf(weather, doy, min_non_zero_days = 20)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{doy}{Day of the year for which the empirical cumulative density function of precipitatio should be calculated}

\item{min_non_zero_days}{minimum amount of non-zero precipitation observation need 
to be present for ecdf calculation, otherwise NA returned}
}
\value{
empirical cumulative density function
}
\description{
Function calculates for a specified day of the year the empirical cumulative density 
function of non-zero precipitation events.
}
\details{
This is a helper function which is mainly used for the spatial corroboration test
of precipitation. The function includes precipitation data of a 29 day window centered at
the day of interest throughout all observed years. Missing observations and 
zer-precipitation observations are discarded. If there are fewer observations than 
\code{min_non_zero_days}, then NA is returned instead of the empirical cumulative
density function.
}
\examples{
get_ecdf(weather = target_weather, doy = 1)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ecdf}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
