% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_metrics.R
\name{get_eval_metrics}
\alias{get_eval_metrics}
\title{Calculate evaluation metrics for weather imputation analysis}
\usage{
get_eval_metrics(
  eval_df,
  eval_fun = c("get_MAE", "RPIQ", "RMSEP", "cor"),
  calc_summary_score = T,
  bigger_better = c(F, T, F, T),
  weights = NA
)
}
\arguments{
\item{eval_df}{data.frame with the columns c(station, patch_method, value and 
original)}

\item{eval_fun}{character vector of the evaluation metric functions which should
be applied. Needs to be exactly the functions name without the parantheses ()}

\item{calc_summary_score}{boolean, decides if summary score is calculated}

\item{bigger_better}{boolean vector, needs to be the same length as 
\code{eval_fun}. Value of TRUE indicates that higher values mean better performance,
FALSE for scores where lower scores indicate better performance. Order needs to
be the same as in \code{eval_fun}}

\item{weights}{numeric vector of same length as \code{eval_fun} and \code{bigger_better}.
If supplied, should contain positive values}
}
\value{
data.frame with station_name, name of evaluation metric and value for
evaluation metric
}
\description{
This function takes the outcomes obtained by \code{\link{get_eval_one_station}}
and calculates several evaluation metrics based on the outcomes.
}
\details{
This is a flexible high-level function similar to 
\code{\link{patch_flexible_several_stations}}. It is compatible with
evaluation metrics supplied with this packaged, but also works with other
packages evaluation metrics or user-defined functions. The only requirements
are the arguments predicted and observed. If the function has differently
named arguments a wrapper function might be necissary.

The function allows also the calculation of a summary score, of the different
evaluation scores. This idea was taken from citation here. In case of negative
values the score rescales the values, so that 0 is the lowest observed 
score, otherwise the score can't handle negative values. It is possible
to assign weights to the individual metrics. If no weights are specified, then
they are equally weighted.

Add citation
}
\examples{
#work on example
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
