% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_one_station.R
\name{get_eval_one_station}
\alias{get_eval_one_station}
\title{Automated comparison and evaluation of several weather patching methods}
\usage{
get_eval_one_station(
  weather,
  weather_info,
  target,
  patch_methods,
  p_missing = 0.3,
  additional_args = NULL,
  method_patches_everything = FALSE,
  period = NULL,
  return_data = "only_new_imputed",
  mute = TRUE
)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{target}{character, column name in weather, on which the method should be applied}

\item{patch_methods, }{character vector with the exact name of the patching 
functions which should evaluated. Needs to be the same length as \code{additional_args}}

\item{p_missing}{numerical, indicates the amount of missing data inserted to
the evaluation period, relative to the evaluation period length. A value of 1 
would remove each observation in the evaluation period, a value of 0 would
lead to no new gaps at all}

\item{additional_args}{by default NULL, if additional arguments of the patching
functions should be specified, then it needs to happen in form of a list
with all the other arguments required by the
patching function. elements of list need to have same name as arguments
of the patching function}

\item{method_patches_everything}{flag, indicates if the method returns all
weather stations or just the target station. This flag is included, so that
patch_flexible_several_stations is also compatible with patching methods
like patch_mice or patch_amelia, which return all patched weather stations
at one function call}

\item{period}{by default NULL, in that case the function searches for the longest
continuous NQA-free period for each weather station. If user-specified
period is wished for, it needs to be a numerical vector of length two, 
row numbers of evaluation period}

\item{return_data}{character, indicates what the function returns after successfull
evaluation. Valid options are "only_new_imputed", which returns only the original and
newly imputed observations per weather station, "evaluation_period" which returns
the whole evaluation period including non-imputed observations and "everything"
which returns a data frame with the same amount as \code{weather}}

\item{mute}{boolean, by default set TRUE. If set FALSE, then function 
keeps user updated on which patching function it is currently working on}
}
\value{
see the description in \code{return_data}
}
\description{
This function allows the evaluation and comparison of several patching methods for the same
weather data.
}
\details{
Evaluation is done by identifying the longest continuous NA-free period in the
target weather data and inserting p_missing amount of NA for desired weather station
in \code{weather}. Targeted weather stations are also used as auxiliary weather stations
along with non-targeted weather stations in \code{weather}.
}
\examples{
get_eval_one_station(weather = weather_Tmin,
weather_info = rbind(target_info, neighbour_info),
target = 'cimis_2', 
patch_methods = c('patch_idw','patch_normal_ratio'), 
method_patches_everything = c(TRUE, FALSE))
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
