% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gap_monthly.R
\name{get_gap_monthly}
\alias{get_gap_monthly}
\title{helper function: detects climatological gaps in monthly weather data}
\usage{
get_gap_monthly(x, temp = TRUE, gap_threshold = 10)
}
\arguments{
\item{x}{numerical vector, usually either daily temperature or precipitation data}

\item{temp}{flag if it is temperature data, in case of precipitation set it as FALSE}

\item{gap_threshold}{testing threshold, if gap at the edge of the distribution
is larger than that, then all values towards the tail of the gap are flagged}
}
\value{
logical vector of the size of x, TRUE indicates that the data was
flagged because of a gap in the ordered data larger than the gap_threshold
}
\description{
The function takes weather data (usually all the observations for a certain month
throughout all years) and checks if there are suspcious large gaps towards the tails
of the distribution.
}
\details{
In case of temperature data both tails of the distribution are evaluated,
taking the median as the starting value. In case of precipitation only the
upper tail is evaluated.

This is a helper function for the higher-level \code{link{perform_gap_check}}
}
\references{
\insertAllCited{}
}
\seealso{
\code{link{perform_gap_check}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
