% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_longterm_mean_and_sd.R
\name{get_longterm_mean_and_sd}
\alias{get_longterm_mean_and_sd}
\title{Helper function for the climate outlier test}
\usage{
get_longterm_mean_and_sd(weather, variable, doy)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable 
in weather}

\item{doy}{day of the year, for which longterm mean and sd should be calculated}
}
\value{
data.frame with columns c(doy, mean, sd)
}
\description{
This function calculates longterm mean and standard deviation of temperature data
for a certain day of interest.
}
\details{
The function uses a a 15-day window centered at the day of interest and uses
each observation throughout all years within that time-window to calculate mean
and standard deviation.

This is a helper function for the higher-level function \code{\link{perform_climate_outlier_check}}
}
\examples{
get_longterm_mean_and_sd(weather = target_weather, variable = "Tmin", doy = 1)
}
\seealso{
\code{\link{perform_climate_outlier_check}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
