% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prec_rank.R
\name{get_prec_rank}
\alias{get_prec_rank}
\title{Calculate the climatological percentile of a certain precipitation event}
\usage{
get_prec_rank(weather, min_non_zero_days = 20)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{min_non_zero_days}{minimum amount of non-zero precipitation observation need 
to be present for ecdf calculation, otherwise NA returned}
}
\value{
vector of same length as rows in weather with percent percentile calculated
as: \code{floor(percentile * 100)}
}
\description{
Function caluclates for a given precipitation event the corresponding climatological
(floored) percentile rank.
}
\details{
This is a helper function which is mainly used for the spatial corrobation test
of precipitation. The function is wrapper of the \code{get_ecdf()} function. 
It includes precipitation data of a 29 day window centered at
the day of interest throughout all observed years and calculates the empiricial
cumulative density function (ecdf). Missing observations and 
zer-precipitation observations are discarded. If there are fewer observations than 
\code{min_non_zero_days}, then NA is returned instead of the ecdf. Then the 
precipitation event of interest is incerted in the ecdf and the corresponding 
percentile rank is calculated. This is done for each precipitation event.
}
\examples{
get_prec_rank(weather = weather)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{ecdf}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
