% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_streaks.R
\name{get_streaks}
\alias{get_streaks}
\title{Identify unusual repetitions in weather data}
\usage{
get_streaks(weather, variable, rep_threshold = 20)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{column name in \code{weather} for which the test is performed. Should
be either Tmin or Tmax. data.frames in \code{aux_list} need to have the same
name}

\item{rep_threshold}{test limit, if streaks this length or greater are found, 
then test returns positive flag}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious. Data which was dropped in the process of the analysis receives
automatically a FALSE
}
\description{
Function to detect unusual amount of repetitions in daily temperature and precipitation
data.
}
\details{
Befor the function searches for streaks, missing data are dropped and in
case of precipitation observations of no precipitation as well. If in 20 (that is 
the default threshold) subsequent days of the remaining data the exact same
value of the variable of interest is noted, then the whole streak is marked by
a flag.

This function is part of the weather quality control scheme after \insertCite{durre_comprehensive_2010;textual}{weatherQC}. For more details 
please refer to section 3 "Basic integrity tests".
}
\examples{
get_streaks(weather = weather, variable = "Tmin")
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
