% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_records.R
\name{get_weather_records}
\alias{get_weather_records}
\title{Download temperature and precipitation records}
\usage{
get_weather_records(region = "world")
}
\arguments{
\item{region}{Character indicating for which level of region the data should be downloaded. Currently only "world" and "USA" are valid options.}
}
\value{
Tibble containing the country-/ subregion-level daily minimum temperature, maximum temperature and precipitation records.
}
\description{
This function scraps the highest and lowest daily temperature (degree C) and in some cases the highest daily precipitation (mm) measurments.
}
\details{
In case of \code{region = "world"} the data will be retrieved from \href{https://en.wikipedia.org/wiki/List_of_countries_and_territories_by_extreme_temperatures}{Wikipedia}. 
The returned tibble contains country-specific temperature records, which might be incomplete. Furthermore, on the country-level
it contains no information on precipitation records.
In case of \code{region = "USA"} the data will be retrieved from the \href{https://www.ncdc.noaa.gov/extremes/scec}{State Climate Extremes Committee (SCEC)}. 
The records are state-wise listed and contain temperature extremes records and precipitation records.
}
\examples{
get_weather_records <- function(region = "world")
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
