% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_frequent_val.R
\name{helper_check_frequent_val}
\alias{helper_check_frequent_val}
\title{Intermediate function for frequent value test}
\usage{
helper_check_frequent_val(val, val_rep, doy, percentile_df)
}
\arguments{
\item{val}{precipitation value which was frequently observed in the weather data}

\item{val_rep}{how often val was observed in a 10 day window}

\item{doy}{day of the year the value was observed}

\item{percentile_df}{data.frame containing the climatological precipitation
percentiles for the corresponding day of the year}
}
\value{
logical, if the frequent value is larger than the corresponding value in 
\code{percentile_df}, then it returns TRUE, otherwise FALSE
}
\description{
Depending on the amount of repetitions the function compares the frequent 
precipitation value with a corresponding climatological percentile of precipitation
for that day of interest.
}
\details{
The more often the value was observed within a 10 day window, the lower the
percentile used for the test. 9 or 10 observations: 30\%, 8 observations: 50\%,
7 observations: 70\%, 5 or 6 observations: 90\%. Climatological precipitation
percentiles are calculated using \code{\link{get_clim_percentiles_prec}}
}
\examples{
#this example is just hypothetical
percentile_df <- get_each_day_precipitation_percentile(weather = weather)
helper_check_frequent_val(val = weather$Precip[2], val_rep = 5,
doy = 2, percentile_df = percentile_df)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
