% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_climatol.R
\name{patch_climatol}
\alias{patch_climatol}
\title{Impute gaps in daily weather data using climatol package}
\usage{
patch_climatol(weather, weather_info, target = NA)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{target}{redundant argument, just needed to be compatible with downstream 
functions}
}
\value{
same data.frame as weather but with all NAs imputed for all columns. There
can be still cases of NA, if for a certain observation none of the other stations
had valid observations
}
\description{
Takes data.frame of daily weather data of several neighboring weather stations
and imputes gaps with iterated normal-ration method of the climatol package.
}
\details{
For more details of the function please refer to \code{\link[climatol]{homogen}}

Climatol requires the weather data via saved objects instead of the workspace
in R and also returns the output as such. In the process of running climatol
this function at first creats the input files in the working directory, retrieves
the climatol from the objects created by climatol and then tries to clean up after
climatol.
}
\examples{
patch_climatol(weather = weather_Tmin, 
              weather_info = rbind(target_info, neighbour_info))
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
