% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_idw.R
\name{patch_idw}
\alias{patch_idw}
\title{Impute gaps in daily weather data using inverse distance weighting}
\usage{
patch_idw(weather, weather_info, phi = 2, target = NULL)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{phi}{distance weighting factor, the higher the factor, the more penalized
far away observations get. By default 2}

\item{target}{actually this argument is not needed and meaningless, but some higher
level function requires it to have this argument}
}
\value{
same data.frame as weather but with all NAs imputed for all columns. There
can be still cases of NA, if for a certain observation none of the other stations
had valid observations
}
\description{
Takes data.frame of daily weather data of several neighboring weather stations
and imputes gaps with the inverse distance weighting (idw) method. The idw-method 
a popular method for example in the missing data imputation of rainfall observations
\insertCite{teegavarapu_improved_2005}{weatherQC}. More weight is assigned to 
the observations of close-by weather stations.
}
\examples{
patch_idw(weather = weather_Tmin, 
weather_info = rbind(target_info, neighbour_info))
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
