% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_mean.R
\name{patch_mean}
\alias{patch_mean}
\title{Impute gaps with mean observation of neighbouring stations}
\usage{
patch_mean(weather, target, weather_info, n_donors = 5, max_dist = 150)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{target}{character, column name in weather, on which the method should be applied}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{n_donors}{amount of auxiliary observations used to fill the gap, by default 5}

\item{max_dist}{maximum distance in km of a auxiliary weather station, to be considered
in the gap imputation}
}
\value{
vector, containing the imputed weather observations of target station.It is
still possible, that cases of NA remain for days none or not enough neighboring stations
had observations available
}
\description{
This function calculates the mean of neighbouring weather stations to fill 
observation gap of the target weather station.
}
\details{
For each day the closest weather station with observations are taken and
the gap is filled with the mean of the auxiliary observations. This is a 
relatively straightforward imputation method and
usually more advanced patching methods yield better results 
\insertCite{addi_evaluation_2022}{weatherQC}. It was nonethless added, to have
a benchmark in patching performance.
}
\examples{
patch_mean(weather = weather_Tmin, 
weather_info = rbind(target_info, neighbour_info),
target = 'cimis_2')
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
