% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_mice.R
\name{patch_mice}
\alias{patch_mice}
\title{Impute gaps in daily weather data using multiple imputation of mice package}
\usage{
patch_mice(
  weather,
  weather_info,
  rain_data = T,
  prcp_threshold = 1,
  max.iter = 5,
  n.impute = 5,
  parallel = F,
  target = NA
)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{rain_data}{logical, in case precipitation data should be patched the 
mice package allows to at first decide if precipitation occures and in a
second step to patch precipitation amount only in cases where the patching
algorithm expects precipitation to happen.}

\item{prcp_threshold}{numerical, setting the threshold until which
precipitation amount the algorithm treats observations are real precipitation
events. Values below are treated as zeros}

\item{max.iter}{maximum amount of iterations of the multiple imputation algorithm,
default is 5}

\item{n.impute}{number of multiple imputations, default is 5}

\item{parallel}{logical, if true the paralleled version of \code{\link[mice]{mice}} 
called \code{\link[micemd]{mice.par}} will be used instead}

\item{target}{redundant argument, just needed to be compatible with downstream 
functions}
}
\value{
same data.frame as weather but with all NAs imputed for all columns. There
can be still cases of NA, if for a certain observation none of the other stations
had valid observations
}
\description{
Takes data.frame of daily weather data of several neighboring weather stations
and imputes gaps with multiple imputation method of the mice-package.
}
\details{
For more details of the function please refer to \code{\link[mice]{mice}}.
Also check out \insertCite{buuren_flexible_2018}{weatherQC} for a more detailed
take on the mice concept and missing data imputation in general.
}
\examples{
patch_mice(weather = weather_Tmin, 
weather_info = rbind(target_info, neighbour_info))
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
