% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_normal_ratio.R
\name{patch_normal_ratio}
\alias{patch_normal_ratio}
\title{Patch daily weather data with the adjusted normal-ratio method}
\usage{
patch_normal_ratio(weather, target, weather_info, n_donors = 5, mute = TRUE)
}
\arguments{
\item{weather}{data.frame with columns for each weather station and rows for
each daily observation. All columns need to contain observations of the same
variable. Missing observations need to be marked with NA}

\item{target}{character, column name in weather, on which the method should be applied}

\item{weather_info}{data.frame containing the name / id of the weather station (needs
to be the same as the column names in weather), Longitude and Latitude in
decimal format}

\item{n_donors}{amount of auxiliary observations used to fill the gap, by default 5}

\item{mute}{boolean, if set TRUE then function can indicate in which cases the patching failed
due to missing neighbouring observations}
}
\value{
vector, containing the imputed weather observations of target station.It is
still possible, that cases of NA remain for days none or not enough neighboring stations
had observations available
}
\description{
Takes data.frame of daily weather data of several neighboring weather stations
and imputes gaps with the adjusted normal-ratio method.
}
\details{
Fleshout the description. Add relevant papers
}
\examples{
#still need to think of examples
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
