% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_climate_outlier_check.R
\name{perform_climate_outlier_check}
\alias{perform_climate_outlier_check}
\title{Test weather data for climatological outliers}
\usage{
perform_climate_outlier_check(
  weather,
  variable,
  max_temperature_z = 6,
  max_prec_threshold = 9,
  max_prec_threshold_freezing = 5,
  prec_percentile = 0.95
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable 
in weather}

\item{max_temperature_z}{maximum size of standardized resiudla of temperature value
to climatological mean for the day of interest}

\item{max_prec_threshold}{factor for test: \code{precipitation >= max_prec_threshold *
precipitation percentile} under non-freezing conditions}

\item{max_prec_threshold_freezing}{factor for test: \code{precipitation >= max_prec_threshold *
precipitation percentile}  under freezing-conditions}

\item{prec_percentile}{precipitation percentile used for the comparison}
}
\value{
Logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
The function takes weather data, calculates long term mean (temperature) or 
percentiles (precipitation) and compares daily weahther observations with it.
}
\details{
In case of temperature long-term mean and standard deviation are calculated using
a 15 day-window centered at the day of interest. All values throughout the observed
years withing that time-window are included in the calculation. In case of precipitation
the percentiles of non-zero preciptiation events are calculated using the function
\code{\link{get_each_day_precipitation_percentile}}

Flags are raised if the standardized residual of temperature exceeds a pre-defined
threshold (by default 6). Precipitation flags are raised if the observation
exceeds the a factor (which changes for freezing conditions) times the percentile 
(by default the 95\% percentile is used).

This function is part of the weather quality control scheme after Durre et al.
(2010) \insertCite{durre_comprehensive_2010}{weatherQC}. For more details 
please refer to section 4 "Outlier checks".
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
