% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_gap_check.R
\name{perform_gap_check}
\alias{perform_gap_check}
\title{Check for large gaps in ordered weather data}
\usage{
perform_gap_check(
  weather,
  variable,
  temp_gap_threshold = 10,
  prec_gap_threshold = 300
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable 
in weather}

\item{temp_gap_threshold}{testing threshold for temperature data, if gaps are equal
or larger, then function flags data. By default 10 degree C}

\item{prec_gap_threshold}{testing threshold for precipitation data, if gaps are equal
or larger, then function flags data. By default 300 mm}
}
\value{
Logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
Takes daily temperature or precipitation data and looks for unusual large gaps in 
ordered data.
}
\details{
The function splits the weather data by month and orders it. In case of temperature
it starts from the median towards both tails of the distribution and searches
for gaps larger than temp_gap_threshold. If such a gap is found, then all the data
coming after the gap towards the end of the tail are flagged. In case of precipitation
only the upper tail is investigated and the start is the first non-zero precipitation
observation. 

This function is part of the weather quality control scheme after \insertCite{durre_comprehensive_2010;textual}{weatherQC}. For more details 
please refer to section 4 "Outlier checks".
}
\examples{
perform_gap_check(weather = target_weather, variable = "Tmin")
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
