% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_target_days.R
\name{select_target_days}
\alias{select_target_days}
\title{Select auxiliary weather station observation for spatial consistency test}
\usage{
select_target_days(df, variable, period_start, period_end)
}
\arguments{
\item{df}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{column name of \code{df} for which data should be retrieved}

\item{period_start}{in date format, indicates start of the period of interest}

\item{period_end}{in date format, indicates end of the period of interest}
}
\value{
numerical vector of the same length as \code{period_start:period_end}
}
\description{
Selects observations of neighboring weather stations to be included in target - 
neighbour weather station linear regression calculation.
}
\details{
Retrieves observations of weather station based on defined time period. In case
parts of the requested period are not present in data.frame, function returns
NA for these days. This function is a helper function for the spatial consistency
test of temperature data.
}
\examples{
select_target_days(weather = target_weather, variable = "Tmin", 
period_start = as.Date("1991-12-17", format = "\%Y-\%m-\%d"), 
period_end = as.Date("1991-02-15", format = "\%y-\%m-\%d"))
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
