% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_consist_one_period.R
\name{spat_consist_one_period}
\alias{spat_consist_one_period}
\title{Test the spatial consistency of target station for one peropd}
\usage{
spat_consist_one_period(
  weather,
  aux_list,
  aux_info,
  period_start,
  variable,
  max_res = 8,
  max_res_norm = 4,
  min_station = 3,
  max_station = 7,
  window_width = 15,
  min_correlation = 0.8,
  min_coverage = 40
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{aux_list}{named list of data.frames with daily weather obsrvations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Strucuture of 
data.frames should be identical of \code{weather}. Data.frames do not necissarily
need to cover excat same time period as \code{weather}}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude")}

\item{period_start}{Date, indicating the start of the month}

\item{variable}{column in \code{weather} for which the test is performed. Should
be either Tmin or Tmax. data.frames in \code{aux_list} need to have the same
name}

\item{max_res}{testing threshold, highest regular resiudal tolerated by test}

\item{max_res_norm}{testing threshold, highest standardized residual tolerated
by the test. Note: both thresholds need to be exceeded in order for the 
test to yield a flag}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{window_width}{amount of extra days added to the target and auxiliary
weather station for the linear regression. Extra days only part of the model
construction, not the testing}

\item{min_correlation}{minimum correlation of model output x target station for 
an auxiliary weather station to be considered in the test}

\item{min_coverage}{minimum amount of shared observations of target and
auxiliary weather stations, so that auxiliary weather station is considered
for the test}
}
\value{
logical vector of same length as target month has days. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
Function checks if target temperature data is in line of neighbouring 
weather stations by linear regression.
}
\details{
The function is a helper function for the higher-level function \code{\link{test_spatial_consistency}}
The function takes temperature data of the target and auxiliary weather station for a
certain month of a certain year plus further observations prior and after the month defined
by the parameter \code{window_width}. By default, target - auxiliary weather stations need to have 
at least \code{min_coverage = 40} shared observations to be considered for linear 
regression. In the next step the index of agreement betwen target and auxiliar weather
stations is calculated. For each value in the target station the closest value of the
each neighbouring series is retrieved using a three day window centered at the
day of interest. The so obtained timeseries of each neighbour station, called
closest_y, is then used for the linear regressions. Weather stations, which 
model linear regression output yielded a correlation score of 0.8 or lower
with actual observations in the target station are dropped. Furthermore,
if more than \code{max_station} auxiliary weather stations are available,
then the ones with the lowest index of agreement are dropped. Weighted mean
of regression output is calculted using the index of agreement as weights. Regular 
and standardized (by subtracting mean residual and deviding by standard
deviation of resiudlas) model residuals are calculated. Model results
within the extrat window period are dropped. Remainin residuals are checked
if they are larger than thresholds defined in \code{max_res} and 
\code{max_res_norm}. Observations for which the regular residual and the 
standardized resiudual both exceed their respective threhsolds are flagged.

For more information please refer to Durre et al. (2010) \insertCite{durre_comprehensive_2010}{weatherQC}
section 6 "Spatial consistency checks" and "Appendix B".
}
\examples{
spat_consist_one_period(weather = weather,
aux_info = aux_info, aux_list = aux_list, 
period_start = as.Date("1992-01-01", format = "\%Y-\%m-\%d"), variable = "Tmin")
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{get_abs_min_difference}}, \code{\link{get_closest_y}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
