% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_info.R
\docType{data}
\name{target_info}
\alias{target_info}
\title{Metadata of target weather station}
\format{
data.frame with 1 row and 6 variables:
\describe{
  \item{id}{unique identifies of weather station}
  \item{Name}{name of weather station}
  \item{Longitude}{longitude of weather station decimal format}
  \item{Latitude}{latitude of weather station in decimal format}
}
}
\source{
\url{https://cimis.water.ca.gov/}
}
\usage{
target_info
}
\description{
data.frames containing a unique identifier for the weather station, coordinates
in decimal format and name of the weather startion.
}
\keyword{datasets}
