% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fixed_limit.R
\name{test_fixed_limit}
\alias{test_fixed_limit}
\title{Test for exceedence of world records}
\usage{
test_fixed_limit(
  weather,
  variable,
  region = NULL,
  subregion = NULL,
  records = NULL
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable in weather}

\item{region}{a character indicating the region for which the records should be downloaded. 
Valid options are `world` and `USA`. Can be also set as \code{NULL} in case user-defined 
limits are supplied}

\item{subregion}{a character, only needed in case \code{region} is defined. In case of 
\code{region = "world"} it should indicate the country the weather data is from. 
In case of \code{region = "USA"} it should be the state's name the weather is obtained from.
If user-defined limits are used, should be set to \code{NULL}}

\item{records}{by default set to \code{NULL}. In case of user-defined limits, it should
be a numeric vector of length two, containing lower and upper limits of tested variable.}
}
\value{
Logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
Wrapper function to test for weather observations outside records of the country / region.
}
\details{
This function tests if observed temperature or precipitation values
exceed the worldwide records. The function also allows for users to 
specify region and subregion, in that case it retrieves the highest and lowest 
temperature (degree C) and daily precipitation (mm) for a specified country / subregion. 
Furthermore, the function also allows user-defined limits. Then it tests which days of the specified variable in the weather data.frame
exceeds the limits and flags these. 

This test is included in the weather quality control schemes \code{durre_qc()};
and \code{costa_qc()}. For more details on the weather quality control functions either 
refer to the help pages of the functions or see \insertCite{costa_gap_2021;textual}{weatherQC}
or \insertCite{durre_comprehensive_2010;textual}{weatherQC}.
}
\examples{
test_fixed_limit(weather = KA_weather, region = 'world', subregion = 'Germany', 
variable = "Tmin")
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{get_weather_records}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
