% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_iterative_temperature_consistency.R
\name{test_iterative_temperature_consistency}
\alias{test_iterative_temperature_consistency}
\title{Check for temporal consistency of temperature variables}
\usage{
test_iterative_temperature_consistency(weather)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}
}
\value{
data.frame with two columns and same amount of rows as in \code{weather}. 
One column is for the flag of minimum temperature, the other for maximum temperature.
Values of \code{TRUE} indicate successful test, meaning that the tested 
variable exceeded the limits of the test.
}
\description{
The function carries out a set of test concerning the temporal consistency of
minimum and maximum temperature. For better testing performance it is adviced
to also supply the average temperature.
}
\details{
The testing algorithm carries out a set of 7 consistency test for each day of the
year. Observations with the highest amount of positive test results are flagged and
the test is carried out until no more new flags are added.

The consistency tests of minimum temperature (Tmin), maximum temperature (Tmax) and
mean temperature (Tmean) of the current day (0) and following day(1) are as followed:
\itemize{
 \item{Test 1: Tmax(0) < Tmin(0) - 1}
 \item{Test 2: Tmean(0) > Tmax(0) + 1}
 \item{Test 3: Tmean(0) < Tmin(0) - 1}
 \item{Test 4: Tmax(0) < Tmin(1) - 1}
 \item{Test 5: Tmin(0) > Tmax(1) + 1}
 \item{Test 6: Tmax(1) < Tmean(0) - 1}
 \item{Test 7: Tmin(1) > Tmean(1) + 1}
}

The + / - 1 tolerance is added for tolerance if different thermometers are used
for the measurement of the variables.

This function is part of the weather quality control scheme after Durre et al.
(2010) \insertCite{durre_comprehensive_2010}{weather_QC}. For more details 
please refer to section 5 "Internal and temporal consistency checks" and "Appendix A".
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
