% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lagged_temperature_range.R
\name{test_lagged_temperature_range}
\alias{test_lagged_temperature_range}
\title{Test lagged temperature range}
\usage{
test_lagged_temperature_range(weather, max_diff = 40)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{max_diff}{threshold for the test, spike and dips lower than this value
are ignored}
}
\value{
data.frame with two columns and same amount of rows as in \code{weather}. 
One column is for the flag of minimum temperature, the other for maximum temperature.
Values of \code{TRUE} indicate successful test, meaning that the tested 
variable exceeded the limits of the test.
}
\description{
The function performs a set of 6 tests regarding the temperature range over
a range of three days per observation.
}
\details{
The test checks if the range of Tmin, Tmax and Tmean over a period of three
days, centered at the day of interest, is larger than a default threshold
of 40 degree C. If so, all values involved in the positive test are flagged.

The set of test include minimum temperature (Tmin), maximum temperature (Tmax) and
mean temperature (Tmean) of the current day (0), previous day (-1) and following
day (1). The tests are as followed:
\itemize{
 \item{Test 1: Tmax(0) >= max[Tmin(-1:1)] + 40}
 \item{Test 1: Tmax(0) >= max[Tmean(-1:1)] + 40}
 \item{Test 1: Tmin(0) <= max[Tmax(-1:1)] - 40}
 \item{Test 1: Tmin(0) <= max[Tmean(-1:1)] - 40}
 \item{Test 1: Tmean(0) <= max[Tmax(-1:1)] - 40}
 \item{Test 1: Tmean(0) >= max[Tmin(-1:1)] + 40}
}


This function is part of the weather quality control scheme after \insertCite{durre_comprehensive_2010;textual}{weatherQC}. For more details 
please refer to section 5 "Internal and temporal consistency checks".
}
\examples{
test_lagged_temperature_range(weather = target_weather)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
