% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_naught_weather.R
\name{test_naught_weather}
\alias{test_naught_weather}
\title{Identify instances of false zeros in weather data}
\usage{
test_naught_weather(weather)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
The function tries to identify instances of false zeros in temperature
data.
}
\details{
The function tests for instances in which both minimum and maximum temperature
received values of 0 degre C or -17.8 degree C (= 0 degree F) but the authors 
more likely intended to record missing data instead. Both variables need to
have 0 or -17.8 as a value for the test to return a flag.

This function is part of the weather quality control scheme after \insertCite{durre_comprehensive_2010;textual}{weatherQC}. For more details 
please refer to section 3 "Basic integrity tests".
}
\examples{
test_naught_weather(weather = weather)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
