% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_precipitation_spatial_corrobation.R
\name{test_precipitation_spatial_corrobation}
\alias{test_precipitation_spatial_corrobation}
\title{Spatial corroboration test of precipitation}
\usage{
test_precipitation_spatial_corrobation(
  weather,
  weather_coords,
  aux_info,
  aux_list,
  max_dist = 75,
  max_station = 7,
  min_station = 3
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{weather_coords}{numerical vector of length two. Should contain longitude and 
latitude (in that order) of target station in decimal format}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude")}

\item{aux_list}{named list of data.frames with daily weather obsrvations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Strucuture of 
data.frames should be identical of \code{weather}. Data.frames do not necissarily
need to cover excat same time period as \code{weather}}

\item{max_dist}{maximum distance in kilometers of neighbouring stations to target station to be 
included in the spatial corrobation test}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
Function to test if the target precipitation is spatial coherent with neighbor
observations
}
\details{
The function compares the precipitation observation of the target station to the
observations of neighboring stations. Test limit depends on the absolute difference
in climatological percentile ranks of precipitation observation. For a detailed
description of the test please refer to \insertCite{durre_comprehensive_2010;textual}{weatherQC}
section 6 "Spatial consistency checks" and "Appendix C". 

Climatological percentile ranks of precipitation is calculated using \code{\link{get_prec_rank}}
}
\examples{
test_precipitation_spatial_corrobation(weather = weather,
weather_coords = c(weather_info$Longitude, weather_info$Latidue),
aux_info = aux_info, aux_list = aux_list)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{get_prec_rank}}, \code{\link{get_abs_min_difference}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
