% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_precipitation_spatial_corroboration.R
\name{test_precipitation_spatial_corroboration}
\alias{test_precipitation_spatial_corroboration}
\title{Spatial corroboration test of precipitation}
\usage{
test_precipitation_spatial_corroboration(
  weather,
  weather_coords,
  aux_info,
  aux_list,
  max_dist = 75,
  max_station = 7,
  min_station = 3
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{weather_coords}{numerical vector of length two. Should contain longitude and 
latitude (in that order) of target station in decimal format}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude")}

\item{aux_list}{named list of data.frames with daily weather obsrvations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Strucuture of 
data.frames should be identical of \code{weather}. Data.frames do not necissarily
need to cover excat same time period as \code{weather}}

\item{max_dist}{maximum distance in kilometers of neighbouring stations to target station to be 
included in the spatial corroboration test}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
Function to test if the target precipitation is spatial coherent with neighbor
observations
}
\details{
The function compares the precipitation observation of the target station to the
observations of neighboring stations. Test limit depends on the absolute difference
in climatological percentile ranks of precipitation observation. For a detailed
description of the test please refer to \insertCite{durre_comprehensive_2010;textual}{weatherQC}
section 6 "Spatial consistency checks" and "Appendix C". 

Climatological percentile ranks of precipitation are calculated using the following approach.
It includes precipitation data of a 29 day window centered at
the day of interest throughout all observed years and calculates the empirical
cumulative density function (ecdf). Missing observations and 
zer-precipitation observations are discarded. If there are fewer observations than 
20 days, then NA is returned instead of the ecdf. Then the 
precipitation event of interest is inserted in the ecdf and the corresponding 
percentile rank is calculated. This is done for each precipitation event.
}
\examples{
test_precipitation_spatial_corroboration(weather = target_weather,
weather_coords = c(target_info$Longitude[1], target_info$Latitude[1]),
aux_info = neighbour_info, aux_list = neighbour_weather)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{test_temperature_corroboration}}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
