% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_temperature_consistency.R
\name{test_temperature_consistency}
\alias{test_temperature_consistency}
\title{Check if Tmin, Tmax and Tmean are in line}
\usage{
test_temperature_consistency(weather, probs = 0.99)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{probs}{numerical indicating the percentile to calculate the test limit}
}
\value{
logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
This function does basic integrity test on minimum, maximum and average daily temperature.
}
\details{
The test composes of four sub-tests, if any of the tests yield positive outcome, 
then the temperature data of that day are flagged as suspicious.


\itemize{
 \item{Test 1: Minimum temperature of day 0 should not be higher than maximum temperature of days -1, 0 and 1}
 \item{Test 2: Minimum temperature of day 0 should not be higher than mean temperature of day -1, 0, or 1}
 \item{Test 3: Maximum temperature of day 0 should not be higher than mean temperature of day -1, 0, or 1}
 \item{Test 4: The absolute difference of reported mean temperature and computed 
 mean temperature (Tmin + Tmax)/2 exceeding the 99\% percentile of all observed values.}
}
}
\examples{
test_temperature_consistency(weather = target_weather)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
