% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_temperature_corroboration.R
\name{test_temperature_corroboration}
\alias{test_temperature_corroboration}
\title{Spatial corroboration test of temperature}
\usage{
test_temperature_corroboration(
  weather,
  weather_coords,
  aux_list,
  aux_info,
  variable,
  max_station = 7,
  min_station = 3,
  max_dist = 75,
  max_diff = 10
)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{weather_coords}{numerical vector of length two. Should contain longitude and 
latitude (in that order) of target station in decimal format}

\item{aux_list}{named list of data.frames with daily weather obsrvations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Strucuture of 
data.frames should be identical of \code{weather}. Data.frames do not necissarily
need to cover excat same time period as \code{weather}}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude")}

\item{variable}{a character indicating the column name of the tested variable 
in weathe}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}

\item{max_dist}{maximum distance in kilometers of neighbouring stations to target station to be 
included in the spatial corroboration test}

\item{max_diff}{maximum difference of the lowest minimum difference of target observation
to neighbouring observations for temperature anomalies}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
Function to test if the target temperature is spatial coherent with neighbour
observations.
}
\details{
The function compares the temperature observation of the target station to the
observations of neighboring stations. This function is a companion test to
\code{\link{test_spatial_consistency}}. While the aim is similar (identify
observations which are not in line with neighboring observations) the testing 
principle is different. While \code{test_spatial_consistency} compares the weighted
output of linear regression to the target value, the \code{test_temperature_corroboration}
compares the target observation to the temperature anomaly of the station to the 
long term mean. The testing tolerance of \code{test_temperature_corroboration} is wider, so it is
less strict, but it also makes fewer demands on the shared observation of 
target and auxiliary weather stations, so it can be applicable where the demands
of \code{test_spatial_consistency} do not permit any test result at all. 

The testing procedure is a follows: The target observation is compared with at
least 3 but never more than 7 neighborliness stations within a 75 km radius.
Neighboring values of the target variable of the same day, but also the previous and
following day are included in the test to account for differences in the recording
protocol. For each observation and station the temperature anomaly is calculated,
meaning the standardized residuals of observation to the long term mean of the day of the year
of the respective weather station. Long term mean and standard deviation are 
calculated using the same principle as in \code{\link{perform_climate_outlier_check}}.
If the minimum absolute difference of the target to neighbor anomaly is larger than 
the default threshold of 10 degree C, then the target value is flagged by the 
test.

For a detailed description of the test please refer to \insertCite{durre_comprehensive_2010;textual}{weatherQC}
section 6 "Spatial consistency checks".
}
\examples{
\dontrun{
test_temperature_corroboration(weather = target_weather,
weather_coords = c(target_info$Longitude, target_info$Latitude),
aux_info = neighbour_info, aux_list = neighbour_weather, variable = "Tmin")
}
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
