% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_temperature_megaconsistency.R
\name{test_temperature_megaconsistency}
\alias{test_temperature_megaconsistency}
\title{Identify temperature inconsistency within calendar months}
\usage{
test_temperature_megaconsistency(weather, min_obs = 140)
}
\arguments{
\item{weather}{data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{min_obs}{minimum amount of observation per calendar month in order 
to be considered in the test, otherwise for the calendar month all observations
automatically get flagged as FALSE}
}
\value{
logical vector of same length as \code{nrow(weather)}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test and is flagged
as suspicious
}
\description{
The function checks if minimum temperature exceeds the highes maximum temperature
of a calendar month or vice versa.
}
\details{
The test is similar to the test regarding internal consistency of temperature.
In the regular consistency test, minimum or maximum observations which did not 
have the accompanying other extreme temperature observation may have slipped
the regular consistency test. The megaconsistency test does not require for
a temperature observation to have the other "patnering" temperature observation 
and thus fill the gap.

This function is part of the weather quality control scheme after Durre et al.
(2010) \insertCite{durre_comprehensive_2010}{weather_QC}. For more details 
please refer to section 7 "Megaconsistency tests".
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
