% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_temporal_continuity.R
\name{test_temporal_continuity}
\alias{test_temporal_continuity}
\title{Test for sudden spikes or dips}
\usage{
test_temporal_continuity(weather, variable, prob = 0.995)
}
\arguments{
\item{weather}{a data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable in weather}

\item{prob}{a numerical indicating the percentile which is used as a threshold for suspicious
observations}
}
\value{
Logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
The function checks for sudden day-to-day jumps or drops in daily weather data.
}
\details{
The decision criterion when strong day-to-day changes in the weather variable 
become suspicious is for this test based on the 99.5\% percentile of all day-to-day
changes in the variable. Every absolute day-to-day change above the percentile
is flagged as suspicious. The function is part of the \insertCite{costa_gap_2021;textual}{weatherQC} quality control algorithm which can be
called via \code{weather_qc_costa}.
}
\examples{
test_temporal_continuity(weather = KA_weather, variable = "Tmin")
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
