% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variable_limit.R
\name{test_variable_limit}
\alias{test_variable_limit}
\title{Test for exceedence of climatological limits}
\usage{
test_variable_limit(weather, variable, probs = c(0.01, 0.99))
}
\arguments{
\item{weather}{a data.frame containing a daily time series data set. 
It should have columns c("Year", "Month", "Day")}

\item{variable}{a character indicating the column name of the tested variable in weather}

\item{probs}{a numeric vector of length two, containing the percentiles used 
to calculate the test threshold. Default is \code{c(0.01, 0.99)}}
}
\value{
Logical vector of same length as rows in \code{weather}. Values of \code{TRUE} indicate successful test,
meaning that the tested variable exceeded the limits of the test.
}
\description{
Checks if daily weather observations or outside the usual climatological ranges of that weather station.
}
\details{
The function calculates monthly and annual percentiles of Tmin and Tmax and 
then tests if the observation exceed the upper and lower percentile.
This function is part of the weather quality control tests of \insertCite{costa_gap_2021;textual}{weatherQC} and is contained by the function 
\code{costa_qc()}
}
\examples{
test_variable_limit(weather = target_weather, variable = "Tmin")
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
