% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_qc_durre.R
\name{weather_qc_durre}
\alias{weather_qc_durre}
\title{Quality control of daily weather data after Durre et al. (2010)}
\usage{
weather_qc_durre(
  weather_list,
  weather_info = NULL,
  aux_list = NULL,
  aux_info = NULL,
  mute = TRUE,
  skip_spatial_test = FALSE,
  duplicate_test_min_obs = 3,
  same_temp_threshold = 10,
  region = NULL,
  subregion = NULL,
  records_temp = NULL,
  records_precip = NULL,
  streak_threshold = 20,
  percentile_frequent_value_test = c(0.3, 0.5, 0.7, 0.9),
  min_non_zero_days = 20,
  temp_gap_threshold = 10,
  prec_gap_threshold = 300,
  max_temperature_z = 6,
  max_prec_threshold = 9,
  max_prec_threshold_freezing = 5,
  prec_percentile_climate_outlier = 0.95,
  dip_threshold = 25,
  lagged_range_max_diff = 40,
  max_dist = 75,
  window_width = 15,
  min_coverage = 40,
  min_correlation = 0.8,
  min_station = 3,
  max_station = 7,
  max_res = 8,
  max_res_norm = 4,
  max_diff_temp_corroboration = 10,
  min_obs_megaconsistency = 140
)
}
\arguments{
\item{weather_list}{list of data.frames containing a daily time series data set. 
It should have columns c("Year", "Month", "Day", "Tmin", "Tmax", "Precip") and
ideally also "Tmean" for full testing power. weather_list elements names should
be the same as the \code{weather_info$id}}

\item{weather_info}{data.frame at least the columns c("id", "Longtitude",
"Latitude"). The column "id" should contain the same names as the named elements of
weather_list. Latitude and longitude should be in decimal format. Number of rows of 
data.frame should be same as number of elements in weather_list}

\item{aux_list}{named list of data.frames with daily weather observations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Structure of 
data.frames should be identical of \code{weather}. Data.frames do not necessarily
need to cover excat same time period as \code{weather}. Can also be set to NULL, in that
case spatial tests are automatically skipped}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude"). Can also be set to NULL, in that
case spatial tests are automatically skipped}

\item{mute}{flag to allow function to communicate testing progress in console
while the test runs}

\item{skip_spatial_test}{flag, which allows to skip the spatial tests because
they can be computationally demanding and may require several hours of runtime}

\item{duplicate_test_min_obs}{minimum amount of non-zero precipitation events
to be considered in the duplicate test. By default it is at least three non-zero
precipitation observations per tested time period (month or year)}

\item{same_temp_threshold}{threhsold for the detection of duplicated temperature
observations}

\item{region}{a character indicating the region for which the records should be downloaded. 
Valid options are `world` and `USA`. Can be also set as \code{NULL} in case user-defined 
limits are supplied}

\item{subregion}{a character, only needed in case \code{region} is defined. In case of 
\code{region = "world"} it should indicate the country the weather data is from. 
In case of \code{region = "USA"} it should be the state's name the weather is obtained from.
If user-defined limits are used, should be set to \code{NULL}}

\item{records_temp}{minimum and maximum temperature in degree Celsius limits for the fixed
limit test. If left NULL, either the function scraps records from the web if}

\item{records_precip}{lower and upper limits of daily precipitation sum (mm) 
for the fixed limit test. If left NULL, either the function scraps records 
from the web if \code{region} and \code{subregion} are supplied, otherwise 
global daily precipitation sum records are used}

\item{streak_threshold}{test limit, if streaks this length or greater are found, 
then test returns positive flag}

\item{percentile_frequent_value_test}{precipitation percentile which will be 
used to to judge if frequently repeated values are to be flagged or not}

\item{min_non_zero_days}{minimum amount of non-zero precipitation observation need 
to be present for ecdf calculation, otherwise NA returned}

\item{temp_gap_threshold}{testing threshold for temperature data, if gaps are equal
or larger, then function flags data. By default 10 degree C}

\item{prec_gap_threshold}{testing threshold for precipitation data, if gaps are equal
or larger, then function flags data. By default 300 mm}

\item{max_temperature_z}{maximum size of standardized resiudla of temperature value
to climatological mean for the day of interest}

\item{max_prec_threshold}{factor for test: \code{precipitation >= max_prec_threshold *
precipitation percentile} under non-freezing conditions}

\item{max_prec_threshold_freezing}{factor for test: \code{precipitation >= max_prec_threshold *
precipitation percentile}  under freezing-conditions}

\item{prec_percentile_climate_outlier}{precipitation percentile used for the comparison}

\item{dip_threshold}{threshold for the test, spike and dips lower than this value
are ignored}

\item{lagged_range_max_diff}{threshold for the test, spike and dips lower than this value
are ignored}

\item{max_dist}{maximum distance in kilometers of neighbouring stations to target station to be 
included in the spatial corroboration test}

\item{window_width}{amount of extra days added to the target and auxiliary
weather station for the linear regression. Extra days only part of the model
construction, not the testing}

\item{min_coverage}{minimum amount of shared observations of target and
auxiliary weather stations, so that auxiliary weather station is considered
for the test}

\item{min_correlation}{minimum correlation of model output x target station for 
an auxiliary weather station to be considered in the test}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{max_res}{testing threshold, highest regular resiudal tolerated by test}

\item{max_res_norm}{testing threshold, highest standardized residual tolerated
by the test. Note: both thresholds need to be exceeded in order for the 
test to yield a flag}

\item{max_diff_temp_corroboration}{maximum difference of the lowest minimum difference of target observation
to neighbouring observations for temperature anomalies}

\item{min_obs_megaconsistency}{minimum amount of observation per calendar month in order 
to be considered in the test, otherwise for the calendar month all observations
automatically get flagged as FALSE}
}
\value{
data.frame with \code{nrow(weather)} rows and the same columns as
in weather, but with six extra columns called c("org_Tmin","org_Tmax", "org_Precip",
"flag_Tmin", "flag_Tmax", "flag_Precip"). In the columns called "org_" plus variable
name the exact copies of observations as in weather can be found, completely unaltered.
In the original column flagged values were replaced with NA. In the same row as where
the original value was removed, a comment indicating which test let to the removal
can be found in the "flag_" plus variable name column.
}
\description{
Performs a series of consistency tests and returns flags of suspicious 
daily weather data.
}
\details{
This weather quality control function was written following a the guidelines of a 
weather control algorithm proposed by \insertCite{durre_comprehensive_2010;textual}{weatherQC}.
In total it consists of 6 groups of tests which contain several independent 
tests:

\itemize{
 \item{Group 1: Basic integrity tests}
    \itemize{
    \item{Naught test \code{\link{test_naught_weather}}}
    \item{Duplicate test \code{\link{get_duplicated_values}}}
    \item{Fixed limit test \code{\link{test_fixed_limit}}}
    \item{Streak test \code{\link{get_streaks}}}
    \item{Frequent value test \code{\link{check_frequent_value}}}
    }
 \item{Group 2: Outlier tests}
 \itemize{
    \item{Gap test \code{\link{perform_gap_check}}}
    \item{Climatological outlier test \code{\link{perform_climate_outlier_check}}}
    }
 \item{Group 3: Internal and temporal consistency tests}
 \itemize{
    \item{Iterative temperature consistency test \code{\link{test_iterative_temperature_consistency}}}
    \item{Spike / dip test \code{\link{test_spike_dip}}}
    \item{Lagged temperature range test \code{\link{test_lagged_temperature_range}}}
    }
 \item{Group 4: Spatial consistency tests}
 \itemize{
    \item{Spatial regression test \code{\link{test_spatial_consistency}}}
    \item{Temperature spatial corroboration test \code{\link{test_temperature_corroboration}}}
    \item{Precipitation spatial corroboration test \code{\link{test_precipitation_spatial_corroboration}}}
    }
 \item{Group 5: Megaconsistency tests}
 \itemize{
    \item{Temperature megaconsistency test \code{\link{test_temperature_megaconsistency}}}
    }
}

The function automatically applies the tests to temperature (Tmin, Tmax) and
Precipitation (Precip) observations within the data.frames of weather_list. The
testing order matters, because flagged observations are removed from the subsequent
tests. This in contrast to \code{\link{weather_qc_costa}} where at least
two tests are needed for an observation to be flagged. The rationale behind the 
testing design in \code{\link{weather_qc_durre}} is that the first tests are
less restrictive but remove the strong outlier (like let's say -9999), which
could reduce the testing quality of more restrictive tests like climatological outlier
test.
}
\examples{
\dontrun{
#prepare input
weather_list <- list(target_weather)
names(weather_list) <- target_info$id

#run without spatial test
weather_qc_durre(weather_list = weather_list, 
skip_spatial_test = TRUE, mute = TRUE)

#run with spatial test 
weather_qc_durre(weather_list = weather_list, 
weather_info = target_info,
aux_list = neighbour_weather,
aux_info = neighbour_info,
mute = TRUE)

#' #run with spatial test and regio-specific test limit
weather_qc_durre(weather_list = weather_list, 
weather_info = target_info,
aux_list = neighbour_weather,
aux_info = neighbour_info,
region = 'USA', subregion = 'California',
mute = TRUE)
}
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
