% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_quantile.R
\name{which_quantile}
\alias{which_quantile}
\title{returns closest value to percentile}
\usage{
which_quantile(x, probs, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{probs}{percentile, which is searched for}

\item{na.rm}{flag, to remove NA values from x}
}
\value{
value in x which is closest to the requested percentile
}
\description{
The function returns the value of a numerical vector which is closest to the
value of the requested percentile. It is the equivalent of \code{\link{which.max}}
or \code{\link{which.min}}.
}
\details{
The function was not written by me but I found it on stakexchange, but I can't find
the respective link anymore.
}
\examples{
which.quantile(x = c(1:4), probs = 0.5)
}
