% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_qc_costa.R
\name{weather_qc_costa}
\alias{weather_qc_costa}
\title{Quality control of daily weather data after Costa et al. (2021)}
\usage{
weather_qc_costa(
  weather_list,
  weather_info = NULL,
  mute = FALSE,
  skip_spatial_test = FALSE,
  aux_list = NULL,
  aux_info = NULL,
  region = NULL,
  subregion = NULL,
  records = NULL,
  probs_variable_limit = c(0.01, 0.99),
  probs_temporal_continuity = 0.995,
  probs_temperature_consistency = 0.99,
  max_dist = 75,
  window_width = 15,
  min_coverage = 40,
  min_correlation = 0.8,
  min_station = 3,
  max_station = 7,
  max_res = 8,
  max_res_norm = 4
)
}
\arguments{
\item{weather_list}{list of data.frames containing a daily time series data set. 
It should have columns c("Year", "Month", "Day", "Tmin", "Tmax", "Precip") and
ideally also "Tmean" for full testing power. weather_list elements names should
be the same as the \code{weather_info$id}}

\item{weather_info}{data.frame at least the columns c("id", "Longtitude",
"Latitude"). The column "id" should contain the same names as the named elements of
weather_list. Latitude and longitude should be in decimal format. Number of rows of 
data.frame should be same as number of elements in weather_list}

\item{mute}{flag to allow function to communicate testing progress in console
while the test runs}

\item{skip_spatial_test}{flag, which allows to skip the spatial tests because
they can be computationally demanding and may require several hours of runtime}

\item{aux_list}{named list of data.frames with daily weather obsrvations of auxiliary
weather stations. Names should be identical to \code{aux_info$id}. Strucuture of 
data.frames should be identical of \code{weather}. Data.frames do not necissarily
need to cover excat same time period as \code{weather}}

\item{aux_info}{data.frame listing the auxiliary weather stations. Should at least contain
the columns c("id", "Longitude", "Latitude")}

\item{region}{a character indicating the region for which the records should be downloaded. 
Valid options are `world` and `USA`. Can be also set as \code{NULL} in case user-defined 
limits are supplied}

\item{subregion}{a character, only needed in case \code{region} is defined. In case of 
\code{region = "world"} it should indicate the country the weather data is from. 
In case of \code{region = "USA"} it should be the state's name the weather is obtained from.
If user-defined limits are used, should be set to \code{NULL}}

\item{records}{by default set to \code{NULL}. In case of user-defined limits, it should
be a numeric vector of length two, containing lower and upper limits of tested variable.
Needs to of length 2, first value indicates lower limit and second the upper limit}

\item{probs_variable_limit}{percentile used as a testing threshold for variable limit test}

\item{probs_temporal_continuity}{percentile uded as a testing threshold for 
temporal conitnuity test}

\item{probs_temperature_consistency}{percentile used as a testing threshold
for temperature consistency test}

\item{max_dist}{maximum distance in kilometers of neighbouring stations to target station to be 
included in the spatial corroboration test}

\item{window_width}{amount of extra days added to the target and auxiliary
weather station for the linear regression. Extra days only part of the model
construction, not the testing}

\item{min_coverage}{minimum amount of shared observations of target and
auxiliary weather stations, so that auxiliary weather station is considered
for the test}

\item{min_correlation}{minimum correlation of model output x target station for 
an auxiliary weather station to be considered in the test}

\item{min_station}{minimum amount of neighbouring stations for the test. If less
is available, then test is not carried out and automatically returns \code{FALSE}
for every observation}

\item{max_station}{maximum number of neighbouring stations included in the test.
If more auxiliary stations available than \code{max_station}, then closest ones
are taken}

\item{max_res}{testing threshold, highest regular resiudal tolerated by test}

\item{max_res_norm}{testing threshold, highest standardized residual tolerated
by the test. Note: both thresholds need to be exceeded in order for the 
test to yield a flag}
}
\value{
data.frame with \code{nrow(weather)} rows and the same columns as
in weather, but with nine extra columns called c("org_Tmin","org_Tmax", "org_Precip",
"flag_Tmin", "flag_Tmax", "flag_Precip", "comment_Tmin", "comment_Tmax", "comment_Precip). 
In the columns called "org_" plus variable name the exact copies of observations 
as in weather can be found, completely unaltered.
In the original column flagged values were replaced with NA. In the same row as where
the original value was removed, a comment indicating which test let to the removal
can be found in the "flag_" plus variable name column. The numbers indicate which
test yielded a positive test. Only in cases of two or more positive tests at
the same time was the flagged observation of the weather variable removed.
}
\description{
Performs a series of consistency tests and returns flags of suspicious data.
}
\details{
This weather quality control function was written following a the guidelines of a 
weather control algorithm proposed by \insertCite{costa_gap_2021;textual}{weatherQC}.
It consists of five consistency tests, namely:

\itemize{
 \item{Test 1: Fixed limit test; see also \code{\link{test_fixed_limit}}}
 \item{Test 2: Variable limit test; see also \code{\link{test_variable_limit}}}
 \item{Test 3: Temporal consistency test; see also \code{\link{test_temporal_continuity}}}
 \item{Test 4: Test for consistency among variables; see also \code{link{test_temperature_consistency}}}
 \item{Test 5: Spatial consistency test; see also \code{\link{test_spatial_consistency}} and \code{\link{test_precipitation_spatial_corroboration}}}
}

The function can be applied to temperature and precipitation data, though the 
test concerning consistency among variables only works for minimum and maximum temperature.
Parts of the test were also described for other weather variables like
relative humidity or wind speed, but this implementation allows only the test
to be used for minimum and maximum temperature and daily precipitation sum. Because
the section regarding spatial consistency was poorly described in the original 
article the spatial consistency test were taken from \insertCite{durre_comprehensive_2010;textual}{weatherQC}.

The main principle of this quality control scheme is, that at least two
positive tests are required for an observation to be flagged. This is intended 
to reduce the amount of false-positive flags.
}
\examples{
weather_list <- list(target_weather)
names(weather_list <- target_info$id)
weather_qc_costa(weather = weather_list, 
mute = T, skip_spatial_test = T)
}
\references{
\insertAllCited{}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
